Maxwell v1.9.2
Over the last month, I've melted the voltage cap zeners
off of the rectifier board. This is not great; they serve
an essential purpose in preventing the mosfets and active
rectifiers behind it from being exposed to an overvoltage
situation and burning out (which has happened to at least
one mosfet).

* Uses much bigger DO-201AE package through-hole diodes.
  These should be able to dissapate a lot more heat, and
  if they do heat up the solder enough to begin melting
  at least they won't be able to fall out of position.
* Adds a P-Ch mosfet between the rectifier and capacitor;
  this should allow me to programmatically turn off the
  dynamo power source when I'm travelling very slowly
  or heading uphill.