Maxwell Remote v1.9.4
It occurred to me that were I to turn off the
regulator, the voltage would just have to be
burned off in the Zener diodes because the dynamo
will just ratchet the voltage up as high as it could.
Really -- I should just use a triac to turn off
the dynamo.

* Adds a (default-enabled) triac for switching
  the dynamo on and off.